Refactor: Improve UI and navigation in login and form fields
Updated the styling of the login title to use `ColorManager.primary`.
Corrected the navigation in the login page "Lets go!" button to navigate to `Routes.homeScreen`.
Renamed the `homeRoute` in `routes.dart` to `homeScreen` for clarity and added the `HomeScreen` to the route mapping.
Styled the custom button with a primary background color and white text.
Adjusted the icon colors in the form container widget to use `ColorManager.secondary` and `ColorManager.primary` for better visual consistency.